graphql-mcp/src/pyproject.toml
[project]
name = "mcp-graphql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "ruff>=0.11.13",
    "aiohttp>=3.11.10",
    "asyncio>=3.4.3",
    "fastmcp>=2.9.2",
]

[tool.ruff]

exclude = [".venv", "__pycache__", ".git", "dist", "build"]
line-length = 120

[tool.ruff.lint]

select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings  
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "UP",  # pyupgrade (modernize code)
]

# Ignore specific rules if needed
ignore = []


[tool.ruff.format]
# Format settings
quote-style = "double"
indent-style = "space"

[tool.uv]
allow-insecure-host = ["mgraphql-mcp"]
